BACKPORT: sched,rt: Remove return value from pull_rt_task()
In order to be able to use pull_rt_task() from a callback, we need to
do away with the return value.

Since the return value indicates if we should reschedule, do this
inside the function. Since not all callers currently do this, this can
increase the number of reschedules due rt balancing.

Too many reschedules is not a correctness issues, too few are.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Cc: ktkhai@parallels.com
Cc: rostedt@goodmis.org
Cc: juri.lelli@gmail.com
Cc: pang.xunlei@linaro.org
Cc: oleg@redhat.com
Cc: wanpeng.li@linux.intel.com
Cc: umgwanakikbuti@gmail.com
Link: http://lkml.kernel.org/r/20150611124742.679002000@infradead.org
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit 8046d6806247088de5725eaf8a2580b29e50ac5a)
Bug: 35833695
Test: Boot
Change-Id: I73024bb9f42ae32a48e208a4a82fc55151a74148
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: khusika <khusikadhamar@gmail.com>