prlmk: Introduce lowmemory notification events
This patch aims to centralize the lowmemory checking mechanism, rather
than duplicating the code across different functions.

The is_low_mem() function will return the appropriate lowmemory
notification, based on which we can run the code relevant to said
lowmemory notification.

There should be no functionality change as a result of this patch.

Signed-off-by: darkhz <kmachanwenw@gmail.com>